// Assets-Uncapped/front-portal-develop/src/services/api/loan-operations/models/BillVendorDTO.ts
export interface BillVendorDTO {
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    creationDate?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    category?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    phoneNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    addressLine1?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    addressLine2?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    region?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    country?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    postcode?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    registrationNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    taxNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountIban?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountSortCode?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountSwiftBic?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountCountry?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountHolderName?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    accountBankName?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    modulrBeneficiaryId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    ccBeneficiaryId?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    status?: BillVendorDTOStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    rejectReason?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    createdBy?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    branchCode?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    bankCode?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    ifscCode?: string;
    /**
     * 
     * @type {string}
     * @memberof BillVendorDTO
     */
    bsbCode?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BillVendorDTO
     */
    local?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BillVendorDTO
     */
    shared?: boolean;
}


/**
 * @export
 */
export const BillVendorDTOStatusEnum = {
    New: 'NEW',
    Approved: 'APPROVED',
    Rejected: 'REJECTED'
} as const;
export type BillVendorDTOStatusEnum = typeof BillVendorDTOStatusEnum[keyof typeof BillVendorDTOStatusEnum];